% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dat_GSE54_GSE48.R
\docType{data}
\name{Dat_GSE54_GSE48}
\alias{Dat_GSE54_GSE48}
\title{Dat_GSE54_GSE48}
\usage{
data(Dat_GSE54_GSE48)
}
\description{
A list containing the data list of raw data and meta data from GSE104154 and GSE161648.
}
\examples{
data(Dat_GSE54_GSE48)
names(Dat_GSE54_GSE48)


}
\references{
Redente et. al. (2020), JCI Insight 6(1); Xie et. al. (2021), Sci Adv 7(24)
}
\keyword{dataset}
