% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Query_transfer.R
\docType{data}
\name{Data_Query_transfer}
\alias{Data_Query_transfer}
\title{Data_Query_transfer}
\usage{
data(Data_Query_transfer)
}
\description{
A list containing the Seurat objects as the query for transfer function. The query data were from GSE161648.
}
\examples{
data(Data_Query_transfer)
names(Data_Query_transfer)


}
\references{
Redente et. al. (2020), JCI Insight 6(1)
}
\keyword{dataset}
