% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Ref_transfer.R
\docType{data}
\name{Data_Ref_transfer}
\alias{Data_Ref_transfer}
\title{Data_Ref_transfer}
\usage{
data(Data_Ref_transfer)
}
\description{
A list containing the Seurat objects as the reference for transfer function. The reference data were obtained from GSE104154.
}
\examples{
data(Data_Ref_transfer)
names(Data_Ref_transfer)


}
\references{
Xie et. al. (2021), Sci Adv 7(24)
}
\keyword{dataset}
