% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_TrData.R
\docType{data}
\name{Example_TrData}
\alias{Example_TrData}
\title{Example_TrData}
\usage{
data(Example_TrData)
}
\description{
A list containing the reference and query Seurat objects. The reference data were obtained from GSE104154 while the query data were from GSE161648.
}
\examples{
data(Example_TrData)
names(Example_TrData)


}
\references{
Redente et. al. (2020), JCI Insight 6(1); Xie et. al. (2021), Sci Adv 7(24)
}
\keyword{dataset}
