% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_TrData_SCT.R
\docType{data}
\name{Example_TrData_SCT}
\alias{Example_TrData_SCT}
\title{Example_TrData_SCT}
\usage{
data(Example_TrData_SCT)
}
\description{
A list containing the reference and query Seurat objects with sctransform normalized. The reference data were obtained from GSE104154 while the query data were from GSE161648.
}
\examples{
data(Example_TrData_SCT)
names(Example_TrData_SCT)


}
\references{
Redente et. al. (2020), JCI Insight 6(1); Xie et. al. (2021), Sci Adv 7(24)
}
\keyword{dataset}
