% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFile.R
\name{ExtractFile}
\alias{ExtractFile}
\title{ExtractFile}
\usage{
ExtractFile(ExtractID = ExtractID, DecompressType = decompressType)
}
\arguments{
\item{ExtractID}{The list of GSE IDs. Required}

\item{DecompressType}{Which types of files to be decompressed if DecompressFile is set to TRUE. Options: "gz"(for .gz),"tar"(for .tar or .tar.gz),"zip"(for .zip) or "All"(for .tar, .tar.gz, .gz, .zip). Optional. Default: "All"}
}
\description{
Decompress the files
}
\examples{

library(singleGEO)
###not run
##myGeoID<-c("GSE131800")
##DownloadFileInfo<-Get_Geo_Data(DownGSEID=myGeoID,DecompressFile=FALSE)
##ExtractFile(ExtractID=myGeoID,DecompressType="tar")
}
