% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Geo_Data.R
\name{Get_Geo_Data}
\alias{Get_Geo_Data}
\title{Get_Geo_Data}
\usage{
Get_Geo_Data(
  DownGSEID = DownGSEID,
  DecompressFile = FALSE,
  DecompressType = "All"
)
}
\arguments{
\item{DownGSEID}{The GSE IDs for data downloading. Required}

\item{DecompressFile}{Whether decompress the files or not. Binary variable with TRUE or FALSE. Optional. Default: FALSE}

\item{DecompressType}{Which types of files to be decompressed if DecompressFile is set to TRUE. Options: "gz"(for .gz),"tar"(for .tar or .tar.gz),"zip"(for .zip) or "All"(for .tar, .tar.gz, .gz, .zip). Optional. Default: "All"}
}
\value{
The list of files downloaded
}
\description{
Download all supplementary data of GSE IDs provided by the user
}
\examples{

library(singleGEO)
###not run
##myGeoID<-c("GSE131800")
##DownloadFileInfo<-Get_Geo_Data(DownGSEID=myGeoID,DecompressFile=FALSE)
}
