% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_GSMFromGSE.R
\name{Get_GSMFromGSE}
\alias{Get_GSMFromGSE}
\title{Get_GSMFromGSE}
\usage{
Get_GSMFromGSE(GeoDataBase = GeoDataBase, GSE.ID = c("GSE122960", "GSE158127"))
}
\arguments{
\item{GeoDataBase}{The object returned by getGeoMetaDatabase function. Required}

\item{GSE.ID}{The GSE IDs}
}
\value{
The GSM information
}
\description{
Get GSM information matching to the GSE IDs
}
\examples{

library(singleGEO)
MyDataBase<-GetGeoMetaDatabase(Sqlfile=NULL,Demo=TRUE)
GSEID<-c("GSE122960","GSE158127")
GSM_info<-Get_GSMFromGSE(GeoDataBase=MyDataBase,GSE.ID=GSEID)

}
