% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Keyword_Meta.R
\name{Get_Keyword_Meta}
\alias{Get_Keyword_Meta}
\title{Get_Keyword_Meta}
\usage{
Get_Keyword_Meta(
  GeoDataBase = GeoDataBase,
  Organism = "Homo sapiens",
  InputSearchList = list(kw1 = "single"),
  Platform = "scRNAseq"
)
}
\arguments{
\item{GeoDataBase}{The object returned by getGeoMetaDatabase function. Required}

\item{Organism}{The organisms to search for. One or multiple organisms can be provided. Optional. Default: c("Homo sapiens")}

\item{InputSearchList}{The list of keywords which are described in GEO database. Multiple keywords can be provided and the keywords should be in a list. Required. Default: c("single")}

\item{Platform}{The platform of single cell sequencing. Currently only support single cell RNAseq(scRNAseq) and single cell ATACseq(scATACseq). Options: scRNAseq, scATACseq. Default: "scRNAseq"}
}
\value{
The filtered meta data
}
\description{
Filter the meta database to obtain the study using single cell RNAseq data and containing the key information provided by user
}
\examples{

library(singleGEO)
MyDataBase<-GetGeoMetaDatabase(Sqlfile=NULL,Demo=TRUE)
MyOrganism<-c("Mus musculus","Homo sapiens")
search_kw1<-c("lung","pulmonary")
search_kw2<-c("fibrosis")
KeyList<-list(kw1=search_kw1,kw2=search_kw2)
MySearch<-Get_Keyword_Meta(GeoDataBase=MyDataBase,Organism=MyOrganism, InputSearchList=KeyList)

}
