% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGeoMetaDatabase.R
\name{GetGeoMetaDatabase}
\alias{GetGeoMetaDatabase}
\title{GetGeoMetaDatabase}
\usage{
GetGeoMetaDatabase(Sqlfile = NULL, Demo = FALSE)
}
\arguments{
\item{Sqlfile}{Using existing GEO meta database or downloading the new one. If "GEOmetadb.sqlite" file has been downloaded but it is not in current folder, the path of the file should be provided. If file of "GEOmetadb.sqlite" can be detected under current folder, the program will use the file. If no "GEOmetadb.sqlite"" file is detected, the program will download the new one and save it in current folder. Optional. Default: NULL(check file under current folder or automatically download the new one)}

\item{Demo}{Whether to use Demo database or not. Default: FALSE}
}
\value{
The meta database
}
\description{
Download and connect to a SQLite database file
}
\examples{

library(singleGEO)
GeoDataBase<-GetGeoMetaDatabase(Sqlfile=NULL,Demo=TRUE)
}
