% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDat_GSE123405_data.R
\docType{data}
\name{listDat_GSE123405_data}
\alias{listDat_GSE123405_data}
\title{listDat_GSE123405_data}
\usage{
data(listDat_GSE123405_data)
}
\description{
A list containing seurat object of two smokers and two non-smoker obtained from GSE123405.
}
\examples{
data(listDat_GSE123405_data)
names(listDat_GSE123405_data)


}
\references{
Zuo et al (2020). Respir Res. 21(1):200
}
\keyword{dataset}
