% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDat_GSE134174.R
\docType{data}
\name{listDat_GSE134174}
\alias{listDat_GSE134174}
\title{listDat_GSE134174}
\usage{
data(listDat_GSE134174)
}
\description{
A list containing seurat object of two heavy smokers and two non-smoker as well as meta data obtained from GSE134174.
}
\examples{
data(listDat_GSE134174)
names(listDat_GSE134174)


}
\references{
Goldfarbmuren et al. (2020) Nat Commun 2020 May 19;11(1):2485.
}
\keyword{dataset}
