% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSeuObj_FromNormalizedRNAData.R
\name{MakeSeuObj_FromNormalizedRNAData}
\alias{MakeSeuObj_FromNormalizedRNAData}
\title{MakeSeuObj_FromNormalizedRNAData}
\usage{
MakeSeuObj_FromNormalizedRNAData(
  NormList = NormList,
  GSE.ID = "Test",
  Frow.which.Norm = "lognorm",
  Feature.selection.method = "vst",
  Nfeatures = 2000,
  Do.scale = TRUE,
  Do.center = TRUE,
  Resolution = 0.8,
  Algorithm = 1,
  Dims = 1:30
)
}
\arguments{
\item{NormList}{A list containing the normalized data. Each element of the list is for one normalized dataset. The name of the element will be used as the ID of the normalized dataset. Required.}

\item{GSE.ID}{The GSE ID for this project. Default: Test}

\item{Frow.which.Norm}{Which normalization method the data are obtained. Two normalization methods are allowed and supported: lognorm (For LogNormalization) and sct(For sctransform). Default: lognorm}

\item{Feature.selection.method}{The method for the top variable feature selection. This will feed to FindVariableFeatures function. Default: vst}

\item{Nfeatures}{The number of top variable features for the FindVariableFeatures function. Default: 2000}

\item{Do.scale}{Whether to scale the data or not. Binary data. This will feed to ScaleData function. Default: TRUE}

\item{Do.center}{Whether to center the data or not. Binary data. This will feed to ScaleData function. Default: TRUE}

\item{Resolution}{The resolution value for FindClusters function. Default: 0.8}

\item{Algorithm}{The algorithm to be used in FindClusters. Default: 1}

\item{Dims}{The number of top dimensions of reduction to use for the functions of FindNeighbors and RunUMAP. Default: 1:30}
}
\value{
A list with each element corresponding to a Seurat object of a normalized data. The object is ready for the downstream analysis by Seurat package.
}
\description{
Make Seurat object from the normalized data which were downloaded from GEO database. To use this function, a list containing the normalized data should be provided. This function is re-written from Seurat package.
}
\examples{

library(singleGEO)

#
data(Norm_GSE134174)
seu_Norm_GSE134174<-MakeSeuObj_FromNormalizedRNAData(NormList=Norm_GSE134174,GSE.ID="GSE134174")
}
