% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSeuObj_FromRawRNAData.R
\name{MakeSeuObj_FromRawRNAData}
\alias{MakeSeuObj_FromRawRNAData}
\title{MakeSeuObj_FromRawRNAData}
\usage{
MakeSeuObj_FromRawRNAData(
  RawList = RawList,
  MtPattern = "^MT-",
  GSE.ID = "Test",
  MinFeature = 200,
  MaxFeature = 7500,
  MinCount = 400,
  MaxCount = 40000,
  MaxMT = 10,
  Norm.method = "lognorm",
  Scale.factor = 10000,
  Feature.selection.method = "vst",
  Nfeatures = 2000,
  Dims = 1:30,
  Npcs =50,
  Resolution = 0.8,
  Algorithm = 1,
  Do.scale = TRUE,
  Do.center = TRUE
)
}
\arguments{
\item{RawList}{A list containing the raw data. Each element of the list is for one raw dataset. The name of the element will be used as the ID of the raw dataset. Required.}

\item{MtPattern}{The pattern to recognize mitochondria gene. Different species can have different labels. For example, for human data, MtPattern='^MT-' is appropriate, while for mouse it should be MtPattern='^mt-'. Optional. Default: MtPattern='^MT-'}

\item{GSE.ID}{The GSE ID for this project. Default: Test}

\item{MinFeature}{Minimal feature count to keep the cell. Default: 200}

\item{MaxFeature}{Maximal feature count to keep the cell. Default: 7500}

\item{MinCount}{Minimal RNA count to keep the cell. Default: 400}

\item{MaxCount}{Maximal RNA count to keep the cell. Default: 40000}

\item{MaxMT}{Percentage of mitochondria gene to filter out the cell. Default: 10}

\item{Norm.method}{The method to normalize data. Two normalization methods are allowed and supported: lognorm (For LogNormalization) and sct(For sctransform). Default: lognorm}

\item{Scale.factor}{The factor to scale up the data. Default: 10000}

\item{Feature.selection.method}{The method for the top variable feature selection. This will feed to FindVariableFeatures function. Default: vst}

\item{Nfeatures}{The number of top variable features for the FindVariableFeatures function. Default: 2000}

\item{Dims}{The number of top dimensions of reduction to use for the functions of FindNeighbors and RunUMAP. Default: 1:30}

\item{Npcs}{The number of pcs to use for the functions of RunPCA. Default: 50}

\item{Resolution}{The resolution value for FindClusters function. Default: 0.8}

\item{Algorithm}{The algorithm to be used in FindClusters. Default: 1}

\item{Do.scale}{Whether to scale the data or not. Binary data. This will feed to ScaleData function. Default: TRUE}

\item{Do.center}{Whether to center the data or not. Binary data. This will feed to ScaleData function. Default: TRUE}
}
\value{
A list with each element corresponding to a Seurat object of a raw data.
}
\description{
Make Seurat object from the raw count data which were downloaded from GEO database. To use this function, a list containing the raw data should be provided. This function is re-written from Seurat package.
}
\examples{

library(singleGEO)

#

data(testData_GSE134174)

test_dat<-testData_GSE134174$TwoRawData
test_meta<-testData_GSE134174$TwoMetaData
list_GSE134174<-Splitdata_MakeDataList(InputData=test_dat,Group=test_meta$Donor)

##The following code setting big number of MaxFeature,MaxCount,MaxMT and small number
##of MinFeature,MinCount will keep all the cells

seu_GSE134174<-MakeSeuObj_FromRawRNAData(RawList=list_GSE134174,GSE.ID="GSE134174",MinFeature=1,MaxFeature=750000,
                                       MinCount=1, MaxCount=4000000,MaxMT=100)
}
