% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuObj_integration.R
\name{SeuObj_integration}
\alias{SeuObj_integration}
\title{SeuObj_integration}
\usage{
SeuObj_integration(
  Object.list = object.list,
  Object.list2 = NULL,
  Frow.which.Norm = "lognorm",
  SampleNameAsReference = NULL,
  NumberOfSampleForReference = 4,
  NumberOfSampleForReference2 = NULL,
  Nfeatures = 2000,
  Do.scale = TRUE,
  Do.center = TRUE,
  Anchor.reduction = "rpca",
  Dims.anchor = 1:50,
  Dims.umap = 1:30,
  K.weight = 100,
  Resolution = 0.8,
  Algorithm = 1,
  Future.globals.maxSize = 12000 * 1024^2
)
}
\arguments{
\item{Object.list}{A list containing the seurat object of each samples. This should be returned by MakeSeuObj_FromNormalizedData or MakeSeuObj_FromRawData. Required.}

\item{Object.list2}{An additional list containing the seurat object of each samples for data integration. This should be returned by MakeSeuObj_FromNormalizedData or MakeSeuObj_FromRawData. Optional: NULL.}

\item{Frow.which.Norm}{Which normalization method the data are obtained. Two normalization methods are allowed and supported: lognorm (For LogNormalization) and sct(For sctransform). Default: lognorm}

\item{SampleNameAsReference}{A string of names to be used as the reference for data integration. The names should match the object names of the input list. Try the command of "names(object.list)" and/or "names(object.list2)" to check. Optional: NULL.}

\item{NumberOfSampleForReference}{The number of objects will be used as the reference during integration. This parameter will be masked if SampleNameAsReference is not NULL. Default: 4}

\item{NumberOfSampleForReference2}{The number of objects in Object.list2 will be used as the reference during integration. If this parameter is null and object.list2 is not null, it is set to be equal of SampleNameAsReference. Default: NULL}

\item{Nfeatures}{The number of top variable features for the SelectIntegrationFeatures function. Default: 2000}

\item{Do.scale}{Whether to scale the data or not. Binary data. This will feed to ScaleData function. Default: TRUE}

\item{Do.center}{Whether to center the data or not. Binary data. This will feed to ScaleData function. Default: TRUE}

\item{Anchor.reduction}{Dimensional reduction method to find anchors. Default: rpca}

\item{Dims.anchor}{The number of top dimensions for the IntegrateData function. Default: 1:50}

\item{Dims.umap}{The number of top dimensions of reduction to use for the functions of FindNeighbors and RunUMAP. Default: 1:30}

\item{K.weight}{Number of neighbors to consider when weighting anchors in "IntegrateData" function. Default: 100}

\item{Resolution}{The resolution value for FindClusters function. Default: 0.8}

\item{Algorithm}{The algorithm to be used in FindClusters. Default: 1}

\item{Future.globals.maxSize}{Maximum allowed total size of global variables. Default: 12000*1024^2 (12Gb)}
}
\value{
A Seurat object with data integrated.
}
\description{
Make Seurat object from the normalized data which were downloaded from GEO database. To use this function, a list containing the normalized data should be provided. This function is re-written from Seurat package.
}
\examples{

library(singleGEO)

#
data(listDat_GSE134174)
listDat_obj<-listDat_GSE134174$data
listDat_meta<-listDat_GSE134174$meta
Int_GSE134174<-SeuObj_integration(Object.list=listDat_obj)
row.names(listDat_meta)<-listDat_meta$Cell
listDat_meta<-listDat_meta[row.names(Int_GSE134174@meta.data),]
Int_GSE134174<-Seurat::AddMetaData(object=Int_GSE134174,metadata =as.data.frame(listDat_meta))
Seurat::DimPlot(Int_GSE134174,group.by="cluster_ident",label=TRUE)
}
