% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splitdata_MakeDataList.R
\name{Splitdata_MakeDataList}
\alias{Splitdata_MakeDataList}
\title{Splitdata_MakeDataList}
\usage{
Splitdata_MakeDataList(InputData, Group = group)
}
\arguments{
\item{InputData}{The input matrix or data frame. Each row is a gene and each column is a cell ID. Required}

\item{Group}{The factor to split the matrix/data frame}
}
\value{
A list of matrix split by Group variable
}
\description{
Split the data by a factor and make a list
}
\examples{

test.matrix<-matrix(sample(0:3,size=900,replace=TRUE),nrow=100,ncol=90)
colnames(test.matrix)<-paste0("Cell",1:90,sep="")
row.names(test.matrix)<-paste0("Gene",1:100,sep="")
sparseMatrix<-as(test.matrix,"sparseMatrix")
pdata<-data.frame(CellID=paste0("Cell",1:90,sep=""),CellGroup=rep(c("A","B","C"),c(20,30,40)))
list_test<-Splitdata_MakeDataList(InputData=sparseMatrix,Group=pdata$CellGroup)
}
