% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData_GSE134174.R
\docType{data}
\name{testData_GSE134174}
\alias{testData_GSE134174}
\title{testData_GSE134174}
\usage{
data(testData_GSE134174)
}
\description{
Single cell RNAseq raw data. Data from two heavy smokers and two non-smokers were extracted from GSE134174.
}
\examples{
data(testData_GSE134174)
names(testData_GSE134174)
}
\references{
Goldfarbmuren et al. (2020) Nat Commun 2020 May 19;11(1):2485.
}
\keyword{datasets}
