% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNormData_GSE134174.R
\docType{data}
\name{testNormData_GSE134174}
\alias{testNormData_GSE134174}
\title{testNormData_GSE134174}
\usage{
data(testNormData_GSE134174)
}
\description{
Single cell RNAseq normalized data. Data from two heavery smokers and two nonsmokers obtained from GSE134174.
}
\examples{
data(testNormData_GSE134174)
names(testNormData_GSE134174)


}
\references{
Goldfarbmuren et al. (2020) Nat Commun 2020 May 19;11(1):2485.
}
\keyword{dataset}
