% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuObj_transfer.R
\name{SeuObj_transfer}
\alias{SeuObj_transfer}
\title{SeuObj_transfer}
\usage{
SeuObj_transfer(
  RefObj = RefObj,
  QueryObj = QueryObj,
  RefLabel = RefLabel,
  Norm.method = "lognorm",
  Future.globals.maxSize = 12000 * 1024^2
)
}
\arguments{
\item{RefObj}{A reference Seurat object. Required.}

\item{QueryObj}{The query Seurat object. Required.}

\item{RefLabel}{The column name of meta data from RefObj. This label will transfer to query data. Required.}

\item{Norm.method}{Normalization method. The standard LogNormalize or sctransform. Options of: "lognorm" or "sct". Default: "lognorm"}

\item{Future.globals.maxSize}{Maximum allowed total size of global variables. Default: 12000*1024^2 (12Gb)}
}
\value{
The Seurat object of query data with predicted labels from reference data.
}
\description{
Transfer reference data information onto query data.
}
\examples{

library(singleGEO)
###

data(Data_Ref_transfer)
data(Data_Query_transfer)

Transfer_Result<-SeuObj_transfer(RefObj=Data_Ref_transfer,
                                QueryObj=Data_Query_transfer,
                                 RefLabel="defined")
Seurat::DimPlot(Transfer_Result,
                group.by="predicted.RefLabel",
                reduction = "umap",
                label=TRUE,
                repel = TRUE)+ NoLegend()
}
